midi_testing/requirements.txt
python-rtmidi>=1.4.9
mido>=1.2.10
pygame>=2.1.0
colorama>=0.4.4
psutil>=5.9.0